+++ lib/hdf5_ch_storage.c
@@ -34,9 +34,9 @@
 static int hdf5_cs_open(struct ch_storage *chns, unsigned flags);
 static int hdf5_cs_close(struct ch_storage *chns);
 static int hdf5_cs_datasync(struct ch_storage *chns);
-static ssize_t hdf5_cs_write(struct ch_storage *chns,
-                             const struct raw_pkt_bsmp*,
-                             size_t);
+static int hdf5_cs_write(struct ch_storage *chns,
+                         const struct raw_pkt_bsmp*,
+                         size_t);
 
 static const struct ch_storage_ops hdf5_ch_storage_ops = {
     .cs_open = hdf5_cs_open,
@@ -181,9 +181,9 @@ static int hdf5_cs_datasync(struct ch_storage *chns)
     return H5Fflush(h5_data(chns)->h5_file, H5F_SCOPE_LOCAL);
 }
 
-static ssize_t hdf5_cs_write(struct ch_storage *chns,
-                             const struct raw_pkt_bsmp *bsamps,
-                             size_t nsamps)
+static int hdf5_cs_write(struct ch_storage *chns,
+                         const struct raw_pkt_bsmp *bsamps,
+                         size_t nsamps)
 {
     /* TODO */
     log_WARNING("XXXXXXXXX "
